+++ include/franka/lowpass_filter.h
@@ -1,0 +1,36 @@
+// Copyright (c) 2017 Franka Emika GmbH
+// Use of this source code is governed by the Apache-2.0 license, see LICENSE
+#pragma once
+
+#include <cmath>
+
+/**
+ * @file lowpass_filter.h
+ * Contains functions for filtering signals with a low-pass filter.
+ */
+
+namespace franka {
+/**
+ * Maximum cutoff frequency
+ */
+constexpr double kMaxCutoffFrequency = 1000.0;
+/**
+ * Default cutoff frequency
+ */
+constexpr double kDefaultCutoffFrequency = 100.0;
+/**
+ * Applies a first-order low-pass filter
+ *
+ * @param[in] sample_time Sample time constant
+ * @param[in] y Current value of the signal to be filtered
+ * @param[in] y_last Value of the signal to be filtered in the previous time step
+ * @param[in] cutoff_frequency Cutoff frequency of the low-pass filter
+ *
+ * @return Filtered value.
+ */
+inline double lowpassFilter(double sample_time, double y, double y_last, double cutoff_frequency) {
+  double gain = sample_time / (sample_time + (1.0 / (2.0 * M_PI * cutoff_frequency)));
+  return gain * y + (1 - gain) * y_last;
+}
+
+}  // namespace franka