+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # CHANGELOG
 
-## 0.4.1 - UNRELEASED
+## 0.5.0 - UNRELEASED
 
- * Fix description of log record entries.
+### Motion and control interfaces
+
+ * **BREAKING** Added low-pass filter for all realtime interfaces with default cutoff frequency
+   of 100 Hz
+ * **DEPRECATED** `franka::Robot::setFilters` in favor of `franka::lowpassFilter`
+ * Fixed description of log record entries
+
+### Library
+
+  * Added `lowpass_filter.h` to public interface
 
 ## 0.4.0 - 2018-06-21